[axl] bugfix: incorrect cast in return from -1

+++ src/axl_io/axl_io_win_File.cpp
@@ -91,7 +91,7 @@ CFile::Read (
 {
 	dword_t ActualSize;
 	bool_t Result = Read (p, (dword_t) Size, &ActualSize, NULL);
-	return Result ? ActualSize : -1;
+	return Result ? (size_t) ActualSize : -1;
 }
 
 size_t
@@ -102,7 +102,7 @@ CFile::Write (
 {
 	dword_t ActualSize;
 	bool_t Result = Write (p, (dword_t) Size, &ActualSize, NULL);
-	return Result ? ActualSize : -1;
+	return Result ? (size_t) ActualSize : -1;
 }
 
 //.............................................................................
